[COS-1611] small poc without extended data and tables
